+++ classes/ui/textUiElement.lua
@@ -32,6 +32,7 @@ function QhunCore.TextUiElement.new(text, settings)
             padding = 0
         }
     )
+    instance._textFrame = nil
 
     -- bind current values
     setmetatable(instance, QhunCore.TextUiElement)
@@ -48,22 +49,41 @@ setmetatable(QhunCore.TextUiElement, {__index = QhunCore.AbstractUiElement})
 -- renders a text element
 function QhunCore.TextUiElement:render(_, parentFrame)
     -- create a font string at the parentFrame
-    local text = parentFrame:CreateFontString()
+    self._textFrame = parentFrame:CreateFontString()
 
-    -- set font, size and other options
-    text:SetFont("Fonts\\FRIZQT__.TTF", self._settings.fontSize)
-    text:SetText(self._text)
-    text:SetJustifyH("LEFT")
-    text:SetTextColor(self._settings.color.r, self._settings.color.g, self._settings.color.b)
+    -- update the view
+    self:update(self._text, self._settings)
 
-    -- add extra padding
-    text._qhunCoreExtraPadding = self._settings.padding
-
-    return text
+    return self._textFrame
 end
 
 -- will be triggered if the storage was reset
 function QhunCore.TextUiElement:onStorageReset()
     -- nothing to do here
     return
+end
+
+-- will be called if the element should be updated
+function QhunCore.TextUiElement:update(text, settings)
+
+    -- apply settings
+    self._text = text
+    self._settings =
+        qhunTableValueOrDefault(
+        settings,
+        {
+            fontSize = 11,
+            color = {r = 1, g = 1, b = 1},
+            padding = 0
+        }
+    )
+
+    -- set font, size and other options
+    self._textFrame:SetFont("Fonts\\FRIZQT__.TTF", self._settings.fontSize)
+    self._textFrame:SetText(self._text)
+    self._textFrame:SetJustifyH("LEFT")
+    self._textFrame:SetTextColor(self._settings.color.r, self._settings.color.g, self._settings.color.b)
+
+    -- add extra padding
+    self._textFrame._qhunCoreExtraPadding = self._settings.padding
 end