+++ changelog.md
@@ -1,9 +1,14 @@
-## 3.0.0 Release
+## 3.0.0 beta.1
 - **Compatible with the game client version 7.3.5**
 - [Feature] Added `QhunCore.ProfileStorage` to easy select between a char and global based storage
 - [Feature] Added `QhunCore.ProfileUiElement` to have a simple tool for managing global and per character settings (needs a profile storage)
 - [Feature] Added `QhunCore.SelectboxUiElement` to create a selectbox in the interface options
+- [Feature] Added `QhunCore.Singleton` as a singleton design pattern abstract class
+- [Feature] Added `QhunCore.Timer` as an interval time related class.
+- [Feature] Added `QhunCore.Interval` as a recurring async task.
+- [Feature] Added `QhunCore.Timeout` as a non recurring async task.
 - [Feature] Added the global function `qhunInstanceOf` to test if an element is a direct instance of a given class
+- [Feature] Added the global function `qhunGetMoneyValue` to easily get the all money values from a total copper value
 - [General] Added the documentation for all features. See `documentation/qhunCore.md` and all subfiles for a complete overview
 
 ## 2.0.0 Release